api/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.9
pyjwt==2.9.0
python-multipart==0.0.9
requests==2.32.3
httpx==0.27.2
sqlalchemy==2.0.35
alembic==1.13.2
python-dotenv==1.0.1
azure-ai-inference==1.0.0b6






